Move date functions to util
